Formatting; Renaming; Documentation
* Formatted files according to existing indentation
* Cleaned up names
* Moved Android classes out of crypto package
* Added javadocs; cleaned up missing docs.
