+++ e3db-crypto-android/src/main/java/com/tozny/e3db/FileSystemManager.java
@@ -18,7 +18,7 @@
  *
  */
 
-package com.tozny.e3db.crypto;
+package com.tozny.e3db;
 
 import android.content.Context;
