+++ e3db-crypto-interface/build.gradle
@@ -1,8 +1,8 @@
 apply plugin: 'java'
 
 dependencies {
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.squareup.okio:okio:1.13.0'
+    compile fileTree(dir: 'libs', include: ['*.jar'])
+    compile 'com.squareup.okio:okio:1.13.0'
 }
 
 sourceCompatibility = "1.7"
