+++ e3db-crypto-plain/build.gradle
@@ -1,9 +1,9 @@
 apply plugin: 'java'
 
 dependencies {
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    compile fileTree(dir: 'libs', include: ['*.jar'])
     compileOnly project(':e3db-crypto-interface')
-    implementation 'org.abstractj.kalium:kalium:0.7.0'
+    compile 'org.abstractj.kalium:kalium:0.7.0'
 }
 
 sourceCompatibility = "1.7"
