+++ plaintest/build.gradle
@@ -7,16 +7,18 @@ compileTestJava {
 }
 
 dependencies {
-    testImplementation("junit:junit:[4,)")
-    testImplementation project(path: ':e3db')
-    testImplementation project(path: ':e3db-crypto-interface')
-    testImplementation project(path: ':e3db-crypto-plain')
-    testImplementation 'com.fasterxml.jackson.core:jackson-core:2.9.0.pr4'
-    testImplementation 'com.fasterxml.jackson.core:jackson-annotations:2.9.0.pr4'
-    testImplementation 'com.fasterxml.jackson.core:jackson-databind:2.9.0.pr4'
+    testCompile("junit:junit:[4,)")
+    testCompile project(path: ':e3db')
+    testCompile project(path: ':e3db-crypto-interface')
+    testCompile(project(path: ':e3db-crypto-plain'))
+    testCompile 'com.fasterxml.jackson.core:jackson-core:2.9.0.pr4'
+    testCompile 'com.fasterxml.jackson.core:jackson-annotations:2.9.0.pr4'
+    testCompile 'com.fasterxml.jackson.core:jackson-databind:2.9.0.pr4'
 }
 
 test {
+    // always re-run tests
+    outputs.upToDateWhen { false }
     testLogging {
         exceptionFormat "full"
         showStackTraces true
