+++ settings.gradle
@@ -2,8 +2,8 @@ include (':e3db',
         ':e3db-crypto-android',
         ':e3db-crypto-plain',
         ':androidtest',
-        ':e3db:e3db-crypto-stub-android',
-        ':e3db:e3db-crypto-stub-plain',
+        ':stubs:e3db-crypto-stub-android',
+        ':stubs:e3db-crypto-stub-plain',
         ':e3db-crypto-interface',
         ':publish',
         ':publish:android',
